refactor(dotnet-sdk): Improve imports for worker usage
