+++ sdk-dotnet/Examples/BasicExample/MyWorker.cs
@@ -1,15 +1,14 @@
-﻿
-using LittleHorse.Worker.Attributes;
+﻿using LittleHorse.Sdk.Worker;
 
 namespace Examples.BasicExample
 {
     public class MyWorker
     {
         [LHTaskMethod("greet-dotnet")]
-        public string Greeting()
+        public string Greeting(string name)
         {
             var message = $"Hello team, This is a Dotnet Worker";
-            Console.WriteLine("Executing task greet: " + message);
+            Console.WriteLine($"Executing task greet {name}: " + message);
             return message;
         }
     }
