--- sdk-dotnet/LittleHorse.Sdk.Tests/Resources/empty_ca.crt
@@ -1,2 +1,0 @@
------BEGIN CERTIFICATE-----
------END CERTIFICATE-----
