--- sdk-dotnet/LittleHorse.Sdk.Tests/Resources/trusted_ca.crt
@@ -1,22 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDuTCCAqGgAwIBAgIUPflYlwU1C8u8yvg+tEYqi2t6bFAwDQYJKoZIhvcNAQEL
-BQAwazELMAkGA1UEBhMCVVMxEzARBgNVBAgMCkNhbGlmb3JuaWExFjAUBgNVBAcM
-DVNhbiBGcmFuY2lzY28xEDAOBgNVBAoMB01vY2sgQ0ExCzAJBgNVBAsMAklUMRAw
-DgYDVQQDDAdtb2NrLWNhMCAXDTI0MDkyNzIzMDc0NFoYDzIxMjQwOTAzMjMwNzQ0
-WjBrMQswCQYDVQQGEwJVUzETMBEGA1UECAwKQ2FsaWZvcm5pYTEWMBQGA1UEBwwN
-U2FuIEZyYW5jaXNjbzEQMA4GA1UECgwHTW9jayBDQTELMAkGA1UECwwCSVQxEDAO
-BgNVBAMMB21vY2stY2EwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQCp
-MuyIDLEkMe1L+NKKgWw9YFExrC4Oy/NGiasslBtfH1PVzS0oA4Ek0H9Sr5JWF/NE
-Gx9CA0W6v7ZEeP29YZis6cZF/CHdbKJVBzzrcj2SOV7/17VoV7NqSxzLw2IoTgxa
-WkUKmmc0XrQS9LxyfTjXjzEgxmt1+9m/M4QwOQli/7moCLD6ZhPlyls1TIzJHRdM
-2C3gPYyzhUxjggWjpGwDCR0mW9g18ZjTqOdWOQYPoom9vpisx2O075IENLx3gRWL
-a4Al+QHyF78LrZ6ZqZqkn4WCP7K5NseXc7kf28fzSbc/19+7tmmQtW1ty/tuDYIe
-QeYIXPDrdUjTPxdtOE2JAgMBAAGjUzBRMB0GA1UdDgQWBBQiHMwtiTq68/lpLgqq
-qaLAEXQU8DAfBgNVHSMEGDAWgBQiHMwtiTq68/lpLgqqqaLAEXQU8DAPBgNVHRMB
-Af8EBTADAQH/MA0GCSqGSIb3DQEBCwUAA4IBAQChPvcwAsiENEj+lpb6IuYSWSBi
-3OK6nHJSkiHe6IaC17kukrL5p6QVYSiVuaADW1W5kvrFfuEPqdRLjlybBvpdIxq0
-Q9/lIuVu77Ic3IkVGJ0j+7dA6Fb1N/Ho+MeDqbfXyYxaHFUiBoL4pir4+j4fIqM8
-oIeLdL08x5H3RIzbg0Y6Y/Fy0wYS9dEK1PzaiS+N+3GlNFCuNYNeUnlfRcNk5kVF
-nJ+qEpxYFDa+tYrPOy2oWg51pQouvoQm6QLajYm0cEA3c4i/RMAYea36PLzkGpM7
-gcvYRwpwL8re9QWyj9A0qEZyJNmkuh3eRCvvA8/6vkzQiLvkMaDKhMFkyJXx
------END CERTIFICATE-----
