+++ sdk-dotnet/LittleHorse.Sdk/Worker/Internal/LHServerConnection.cs
@@ -1,10 +1,9 @@
 ﻿using Grpc.Core;
-using LittleHorse.Worker.Internal.Helpers;
 using LittleHorse.Common.Proto;
+using LittleHorse.Sdk.Helper;
 using Microsoft.Extensions.Logging;
-using LittleHorse.Sdk.Internal;
 
-namespace LittleHorse.Worker.Internal
+namespace LittleHorse.Sdk.Worker.Internal
 {
     public class LHServerConnection<T> : IDisposable
     {
