+++ sdk-dotnet/LittleHorse.Sdk/Worker/VariableMapping.cs
@@ -1,10 +1,9 @@
-﻿using LittleHorse.Common.Exceptions;
-using LittleHorse.Worker.Internal.Helpers;
+﻿using LittleHorse.Common.Proto;
+using LittleHorse.Sdk.Exceptions;
+using LittleHorse.Sdk.Helper;
 using Microsoft.Extensions.Logging;
-using LittleHorse.Common.Proto;
-using LittleHorse.Sdk.Internal;
 
-namespace LittleHorse.Worker.Internal
+namespace LittleHorse.Sdk.Worker
 {
     public class VariableMapping
     {
